add auth to delte and update

+++ app/Http/Controllers/ListingController.php
@@ -45,7 +45,7 @@ class ListingController extends Controller
         }
 
         $formFields['user_id'] = auth()->id();
-        
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -60,6 +60,11 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+        // make sure logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorazed Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -82,6 +87,10 @@ class ListingController extends Controller
     // Delete Listing
     public function destroy(Listing $listing)
     {
+        // make sure logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorazed Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
